new test file set up. minor changes to make two array functions not have hardcoded text, so they can be resued easier
