+++ apps/core/views/monthly_check_in.py
@@ -66,7 +66,7 @@ def save_check_in_transactions(request, transaction_array):
             new_transaction.bucket = Bucket.objects.get(id=transaction["bucket_id"])
             new_transaction.amount = transaction["amount"]
             new_transaction.transactionDate = timezone.now()
-            new_transaction.description = 'Check-In adjustment'
+            new_transaction.description = transaction["description"]
             new_transaction.save()
 
 
@@ -75,7 +75,7 @@ def save_account_status(account_status_array):
             account_status = AddBankAccountStatus().save(commit=False)
             account_status.amount = account["amount"]
             account_status.status_date = account["date"]
-            account_status.description = 'hardcoded description text. You have no control!'
+            account_status.description = account["description"]
             current_bank_account = BankAccount.objects.get(id=int(account["account_id"]))
             account_status.bank_account = current_bank_account
             account_status.save()
@@ -150,7 +150,8 @@ def create_array_from_form(request, post_type_array):
             input_amount = 0
         new_array.append({
             'bucket_id': item,
-            'amount': input_amount
+            'amount': input_amount,
+            'description': 'Check-In adjustment',
         })
     return new_array
 
@@ -171,6 +172,7 @@ def create_account_status_array(request):
         account_status_array.append({
             'account_id': item,
             'date': request.POST['date__' + item],
-            'amount': request.POST['amount__' + item]
+            'amount': request.POST['amount__' + item],
+            'description': 'Account value as of this date'
         })
     return account_status_array
